Initial Fork
csoM modifications.

Including:
1. Show and modify Telldus Events/Schedules.
2. Get alerts when device changes state.
3. Get push notifications.
etc….

+++ inc/settings_users.php
@@ -1,4 +1,12 @@
 <?php
+
+
+	/* Check access
+	--------------------------------------------------------------------------- */
+	if ($user['admin'] != 1) {
+			header("Location: ?page=settings&view=user&action=edit&id={$user['user_id']}");
+			exit();
+	}
 	
 	echo "<h3>".$lang['Users']."</h3>";
 
@@ -24,6 +32,7 @@
 				echo "<th>#</th>";
 				echo "<th>".$lang['Email']."</th>";
 				echo "<th>".$lang['Admin']."</th>";
+				if ($config['log_activity'] != 0) echo "<th>".$lang['Last_active']."</th>";
 				echo "<th></th>";
 			echo "</tr>";
 		echo "</thead>";
@@ -41,10 +50,16 @@
 					echo "<td>";
 						if ($row['admin'] == 1) echo "<img style='height:16px;' src='images/metro_black/check.png' alt='icon' />";
 					echo "</td>";
+					
+					if ($config['log_activity'] != 0) {
+						echo "<td>";
+								if ($row['last_active'] != 0) echo date("Y-m-d H:i:s", $row['last_active']);
+						echo "</td>";
+					}
 
 					echo "<td style='text-align:right;'>";
+						if ($_SESSION['fuTelldus_user_loggedin'] != $row['user_id']) echo "<a class='btn btn-danger' href='?page=settings_exec&action=userDelete&id={$row['user_id']}' onclick=\"return confirm('".$lang['Are you sure you want to delete']."')\">".$lang['Delete']."</a>";
 						echo "<a class='btn' href='?page=settings&view=user&action=edit&id={$row['user_id']}'>".$lang['Edit']."</a> &nbsp; ";
-						echo "<a class='btn btn-danger' href='?page=settings_exec&action=userDelete&id={$row['user_id']}' onclick=\"return confirm('".$lang['Are you sure you want to delete']."')\">".$lang['Delete']."</a>";
 					echo "</td>";
 
 				echo "</tr>";
